+++ static/bundle.js
@@ -108,8 +108,12 @@
     });
 
     clearButton.addEventListener("click", () => {
-        localStorage.clear();
-        location.reload();
+        const warningClearStoredData = "Make sure you have saved or submitted your data before proceeding.\n\n" +
+            "Delete forever?";
+        if (confirm(warningClearStoredData) == true) {
+            localStorage.clear();
+            location.reload();
+        }
     });
 
     function getHourAndMinute() {
